+++ utils/__init__.py
@@ -1,5 +1,5 @@
 from utils.nodes import list_node
-from utils.addRecordAliDNS import AliDns
-from utils.addRecordCloudflare import CloudflareDns
+from utils.alidns import AliDns
+from utils.cloudflare import CloudflareDns
 from utils.base import BaseDns
 # 这东西也需要注释吗？
